Finished v.01

+++ src/common.c
@@ -1,15 +1,15 @@
 #include "common.h"
 
-i64 *indlist(i64 n) {
+i32 *indlist(i32 n) {
   srand(0);
-  i64 *ret=(i64*)malloc(sizeof(i64)*n);
-  i64 i;
+  i32 *ret=(i32*)malloc(sizeof(i32)*n);
+  i32 i;
   for (i=0;i<n;ret[i]=i,i++);
   return ret;
 }
 
-void shuffle_inds(i64* arr, i64 size) {
-  i64 i,j,tmp;
+void shuffle_inds(i32* arr, i32 size) {
+  i32 i,j,tmp;
   for (i=0;i<size;i++) {
     j=rand()%size;
     tmp=arr[j];
